moved loading to component and fix scrollview list to flatlist
