+++ src/screens/post/modal/city.js
@@ -1,15 +1,14 @@
-import {View} from "react-native"
+import {FlatList, View} from "react-native"
 import React, {useEffect, useState} from "react"
 import {StyleSheet} from "react-native"
 import HeaderModal from "@components/headerModal/index"
 import SearchBar from "@components/searchbar"
 import ProvinceItem from "./container/provinceItem"
-import {ScrollView} from "react-native"
 import {useTranslation} from "react-i18next"
-import Loading from "./container/loading"
 import {getAllAddress} from "@utils/address"
 import {useNavigation} from "@react-navigation/native"
 import {nonAccentVietnamese} from "@utils/nonAccentVietnamese"
+import Loading from "@components/loading/index"
 
 const City = () => {
   const [code, setCode] = useState(null)
@@ -36,6 +35,10 @@ const City = () => {
     setSearch(text)
   }
 
+  const renderCityItem = ({item}) => (
+    <ProvinceItem code={code} setCode={setCode} item={item} key={item.code} />
+  )
+
   return (
     <View style={styles.container}>
       <HeaderModal onClose={onGoBack} title={t("post:selectCity")} />
@@ -47,16 +50,11 @@ const City = () => {
       {refreshing ? (
         <Loading />
       ) : (
-        <ScrollView>
-          {data.map((item, index) => (
-            <ProvinceItem
-              code={code}
-              setCode={setCode}
-              item={item}
-              key={item.code}
-            />
-          ))}
-        </ScrollView>
+        <FlatList
+          data={data}
+          renderItem={renderCityItem}
+          keyExtractor={(item) => item.code}
+        />
       )}
     </View>
   )
