--- src/screens/post/modal/container/district/loading.js
@@ -1,22 +1,0 @@
-import {View, ActivityIndicator, StyleSheet} from "react-native"
-import React from "react"
-import Colors from "@common/Color"
-
-const Loading = () => {
-  return (
-    <View style={styles.container}>
-      <ActivityIndicator size="large" color={Colors.orange} />
-    </View>
-  )
-}
-
-export default Loading
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "white",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-})
